+++ R/create_drake_spec.R
@@ -21,21 +21,9 @@ create_drake_spec <- function(
   )
   imports <- cds_prepare_imports(args)
   imports_kernel <- cds_imports_kernel(args, imports)
-  import_spec <- memo_expr(
-    cds_analyze_imports(args, imports),
-    args$cache,
-    imports_kernel
-  )
+  import_spec <- cds_analyze_imports(args, imports)
   knitr_hash <- cds_get_knitr_hash(args)
-  command_spec <- memo_expr(
-    cds_analyze_commands(args),
-    args$cache,
-    args$plan,
-    args$trigger,
-    import_spec,
-    imports_kernel,
-    knitr_hash
-  )
+  command_spec <- cds_analyze_commands(args)
   cds_set_knitr_files(args = args, spec = command_spec)
   out <- c(import_spec, command_spec)
   list2env(out, parent = emptyenv(), hash = TRUE)
@@ -295,6 +283,7 @@ cds_dynamic_deps <- function(dynamic, target, args) {
   UseMethod("cds_dynamic_deps")
 }
 
+#' @export
 cds_dynamic_deps.dynamic <- function(dynamic, target, args) {
   dynamic$.trace <- NULL
   out <- ht_filter(args$ht_globals, all.vars(dynamic))
@@ -307,6 +296,7 @@ cds_dynamic_deps.dynamic <- function(dynamic, target, args) {
   out
 }
 
+#' @export
 cds_dynamic_deps.default <- function(dynamic, target, args) {
   character(0)
 }
@@ -315,10 +305,12 @@ cds_dynamic_trace <- function(dynamic, args) {
   UseMethod("cds_dynamic_trace")
 }
 
+#' @export
 cds_dynamic_trace.dynamic <- function(dynamic, args) {
   all.vars(dynamic$.trace)
 }
 
+#' @export
 cds_dynamic_trace.default <- function(dynamic, args) {
   character(0)
 }
@@ -327,6 +319,7 @@ cds_assert_trace <- function(dynamic, spec) {
   UseMethod("cds_assert_trace")
 }
 
+#' @export
 cds_assert_trace.group <- function(dynamic, spec) {
   bad <- setdiff(spec$deps_dynamic_trace, spec$deps_dynamic)
   if (!length(bad)) {
@@ -343,6 +336,7 @@ cds_assert_trace.group <- function(dynamic, spec) {
   )
 }
 
+#' @export
 cds_assert_trace.dynamic <- function(dynamic, spec) {
   bad <- setdiff(spec$deps_dynamic_trace, spec$deps_dynamic)
   if (!length(bad)) {
@@ -360,6 +354,7 @@ cds_assert_trace.dynamic <- function(dynamic, spec) {
   )
 }
 
+#' @export
 cds_assert_trace.default <- function(dynamic, spec) {
   character(0)
 }