+++ tests/testthat/test-9-clustermq.R
@@ -1,6 +1,7 @@
 drake_context("clustermq")
 
 test_with_dir("clustermq parallelism for CRAN", {
+  skip_on_cran()
   skip_if_not_installed("clustermq", minimum_version = "0.9.1")
   skip_on_os("windows")
   options(clustermq.scheduler = "multicore")
@@ -45,8 +46,7 @@ test_with_dir("clustermq parallelism", {
       caching = caching,
       envir = e,
       verbose = 1L,
-      garbage_collection = TRUE,
-      lock_envir = TRUE
+      garbage_collection = TRUE
     )
     expect_equal(outdated_impl(config), character(0))
     make(
@@ -55,8 +55,7 @@ test_with_dir("clustermq parallelism", {
       jobs = jobs,
       caching = caching,
       envir = e,
-      verbose = 1L,
-      lock_envir = TRUE
+      verbose = 1L
     )
     expect_equal(justbuilt(config), character(0))
     e$my_plan$command[[2]] <- as.call(
@@ -68,8 +67,7 @@ test_with_dir("clustermq parallelism", {
       jobs = jobs,
       caching = caching,
       envir = e,
-      verbose = 1L,
-      lock_envir = TRUE
+      verbose = 1L
     )
     expect_equal(justbuilt(config), "small")
     clean(small, cache = config$cache)
@@ -79,8 +77,7 @@ test_with_dir("clustermq parallelism", {
       jobs = jobs,
       caching = caching,
       envir = e,
-      verbose = 1L,
-      lock_envir = TRUE
+      verbose = 1L
     )
     expect_equal(justbuilt(config), "small")
   }
@@ -172,8 +169,7 @@ test_with_dir("Start off with non-HPC targets, then go to HPC targets.", {
     jobs = jobs,
     envir = e,
     verbose = 1L,
-    garbage_collection = TRUE,
-    lock_envir = TRUE
+    garbage_collection = TRUE
   )
   config <- drake_config(e$my_plan, envir = e)
   expect_equal(sort(justbuilt(config)), sort(e$my_plan$target))
@@ -184,6 +180,7 @@ test_with_dir("Start off with non-HPC targets, then go to HPC targets.", {
 })
 
 test_with_dir("clustermq warnings (worker caching)", {
+  skip_on_cran()
   build <- list(target = "x", warnings = "y")
   expect_warning(cmq_report_warnings(build))
 })